feat: Refactor components and main.py
