+++ docs/src/content/docs/es/hybrid/mves/kafka.mdx
@@ -41,20 +41,20 @@ Esta guía te muestra cómo trabajar con **Apache Kafka** localmente usando **Do
 2. **Ejecutar el ejemplo**:
    <Tabs>
       <TabItem label="Python">
-         Produce los eventos y vuelve a consumirlos en una sola ejecución:
-         ```bash
-         python main.py
-         ```
-      </TabItem>
-      <TabItem label="Live consumer">
-         1. **Ejecuta** el consumidor de larga duración en un terminal:
-            ```bash
-            scripts/run_consumer.sh
-            ```
-         2. **Ejecuta** el productor en un segundo terminal y observa cómo llegan los eventos:
-            ```bash
-            scripts/run_main.sh
-            ```
+         - **Ejecuta** el ciclo completo. Publica los eventos, los vuelve a consumir y termina:
+           ```bash
+           python main.py
+           ```
+         - **Divídelo** en dos terminales. `-c` mantiene el consumidor a la escucha hasta `CTRL+C`, y `-p` solo publica:
+
+           Terminal 1:
+           ```bash
+           python main.py -c
+           ```
+           Terminal 2:
+           ```bash
+           python main.py -p
+           ```
       </TabItem>
       <TabItem label="Kafka CLI">
          1. **Produce** un evento escribiéndolo en el console producer:
@@ -76,24 +76,15 @@ Esta guía te muestra cómo trabajar con **Apache Kafka** localmente usando **Do
 
 </Steps>
 
-:::note[Consumer groups]
-`main.py` y `consume.py` comparten el `KAFKA_GROUP_ID` definido en `.env`, por lo que cada ejecución continúa desde el último offset confirmado en vez de releer el topic completo.
-:::
-
 ## Cómo depurar
 
 1. **main.py**:
    - **Abre**: Abre `main.py`.
-   - **Breakpoints**: Añade breakpoints en el código.
-   - **Ejecuta**: Pulsa `F5` y selecciona **Python: Main**.
+   - **Breakpoints**: Añade breakpoints en `publish_events` o en `print_event`.
+   - **Ejecuta**: Pulsa `F5` y selecciona **Python: Main**, que ejecuta el ciclo completo.
 
-2. **consume.py**:
-   - **Abre**: Abre `consume.py`.
-   - **Breakpoints**: Añade breakpoints en el callback del consumidor.
-   - **Ejecuta**: Pulsa `F5` y selecciona **Python: Consumer**, y luego produce eventos desde otro terminal.
-
-3. **Tests**:
-   - **Abre**: Abre un archivo de test (p. ej., `tests/providers/test_event_consumer.py`).
+2. **Tests**:
+   - **Abre**: Abre un archivo de test (p. ej., `tests/components/test_event_consumer.py`).
    - **Breakpoints**: Añade breakpoints en el código de test.
    - **Ejecuta**: Usa la pestaña **Testing** de VS Code y haz clic en el icono **Debug Test** junto al test que quieras depurar.
 
@@ -104,8 +95,6 @@ Esta guía te muestra cómo trabajar con **Apache Kafka** localmente usando **Do
   scripts/run_tests.sh
   ```
 - **Individualmente**: Usa la pestaña **Testing** de VS Code para ejecutar o depurar casos de test específicos.
-
-Los tests usan el topic y el consumer group dedicados definidos en `.env.test`, creados y eliminados en cada test, de forma que nunca tocan el topic que usa `main.py`.
 
 ## Validar resultados
