--- src/hybrid/mves/kafka/scripts/run_consumer.sh
@@ -1,4 +1,0 @@
-#!/bin/bash
-set -e
-
-.venv/bin/python consume.py
